♻️ refactor/ 중복 로직 유틸화 -> 슈퍼베이스에서 users_note fetch로직을 유틸화 했습니다
